+++ pytest.ini
@@ -137,7 +137,7 @@ flake8-ignore =
     medusa/providers/torrent/xml/__init__.py D104
     medusa/providers/torrent/xml/kat.py D100 D102 D202 D204 D400 E501
     medusa/providers/torrent/xml/torrentz2.py D100 D102 D202 D204 D400
-    medusa/rssfeeds.py D100 D103 N802
+    medusa/rss_feeds.py D100 D103 N802
     medusa/sab.py D100 D202 D400 D401 N802 N806
     medusa/sbdatetime.py D100 D101 D102 D105 D202 D205 D400 N801
     medusa/scene_exceptions.py D100
